Booking Service Added and Rental Service Updated
