+++ VacationRental.Api/Controllers/RentalsController.cs
@@ -11,13 +11,10 @@ namespace VacationRental.Api.Controllers
     [ApiController]
     public class RentalsController : ControllerBase
     {
-        private readonly IDictionary<int, RentalViewModel> _rentals;
-
         private readonly IRentalService _rentalService;
 
-        public RentalsController(IDictionary<int, RentalViewModel> rentals, IRentalService rentalService)
+        public RentalsController(IRentalService rentalService)
         {
-            _rentals = rentals;
             _rentalService = rentalService;
         }
 
@@ -25,23 +22,12 @@ namespace VacationRental.Api.Controllers
         [Route("{rentalId:int}")]
         public RentalViewModel Get(int rentalId)
         {
-            if (!_rentals.ContainsKey(rentalId))
-                throw new ApplicationException("Rental not found");
-
-            return _rentals[rentalId];
+            return _rentalService.GetRental(rentalId);
         }
 
         [HttpPost]
         public ResourceIdViewModel Post(RentalBindingModel model)
-        {
-            //var key = new ResourceIdViewModel { Id = _rentals.Keys.Count + 1 };
-
-            //_rentals.Add(key.Id, new RentalViewModel
-            //{
-            //    Id = key.Id,
-            //    Units = model.Units,
-            //    PreparationTimeInDays = model.PreparationTimeInDays
-            //});
+        {           
             return _rentalService.CreateRental(model);
         }
     }
